Speed up access list member counts. (#3545)
* Speed up access list member counts.

Access list member counts will be sped up by proper use of the new access list
cache, as well as by calculating the member counts in the backend as opposed
to having the UI perform the count manually. This will significantly cut down
on the number of calls being sent over gRPC.

* Fix test.

Source-commit: gravitational/teleport.e@bdc3015fe38330768507fc2c0e14a6dfa10f1d49
Source-pull-requests: gravitational/teleport.e#3545

+++ e/lib/web/accesslist_test.go
@@ -85,7 +85,7 @@ func TestGetAccessLists(t *testing.T) {
 	var accessListResp ui.AccessListsResponse
 	require.NoError(t, json.Unmarshal(resp.Bytes(), &accessListResp))
 	require.Len(t, accessListResp.AccessLists, 2)
-	require.Equal(t, 1, *accessListResp.AccessLists[0].MembersCount)
+	require.Equal(t, uint32(1), *accessListResp.AccessLists[0].MembersCount)
 
 	require.Empty(t, cmp.Diff(
 		[]*accesslist.AccessList{accessListResp.AccessLists[0].AccessList, accessListResp.AccessLists[1].AccessList},
